phase2/phase2A/requirements.txt
# Phase 2A: Modular ADK Architecture Dependencies

# Core ADK Framework
google-adk>=1.0.0

# Message Bus and Communication
redis>=4.5.0
aioredis>=2.0.0

# Data Validation and Serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async Support
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Utilities
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Circuit Breaker Pattern
pybreaker>=1.0.0

# Health Checks
healthcheck>=1.3.0

# Performance Monitoring
psutil>=5.9.0
